moved calculations out of interactor
